adding AddressTranslator interface and impl for use in ec2
This change introduces the AddressTranslator interface, which is
intended for use during discovery and node-event processing to translate
peer addresses of nodes. The primary use -- which is driving the
change -- is to translate public IPs to private IPs (when possible)
in ec2.

This solution is common among other CQL driver implementations. The
specific implementation here also follows the convention set by
HostFilter.

We try our best to have test coverage for the bits that changed. Missing
are:

* unit tests for EC2MultiRegionAddressTranslator (an integration test exists)
* unit tests for translation in GetHosts()

Signed-off-by: Justin "Gus" Knowlden <gus@gusg.us>